refactor state
Moves away from event-based history, as it proved to have performance issues on the long run when time-travelling.

Now it keeps the history by storing the whole state every time, but using an immutable pattern which keeps memory somewhat low.

Probably we will need to add a limit on history length, by removing the oldest values.