fix stream cross-references when using async operators
The problem was as follows:

viewConfig = tradingAccount.pipe(...)
positionsMap = tradingAccount.pipe(
  switchMap(() => EMPTY)
);

There was a logic in place that was only meant for concat/switchMap, in which when a stream sends a `next`, we look on what new subscriptions is causing so we can wire up the internal observable.

In this case, the idea was that when tradingAccount emits, switchMap will syncronously subscribe to `EMPTY`, so now we can know that positionsMap also depends on EMPTY.

The problem was that we don't have a direct reference to the dependants of switchMap when the subscription on the inner streamhappens, so we pushed all the dependants of tradingAccount to the observable stack. We were assuming that all dependants were from that switchMap branch, but that's not true.

In this example, all the dependants of tradingAccount were positionsMap *and* viewConfig, so now we were drawing a dependency from EMPTY to the both of them (when it should only be to positionsMap.

The solution I managed to do is if we focus on `next`s instead of subscriptions. The idea is that we can draw a dependency on an observable that emits next to the one that synchronously receives it.

====

This fix was not enough, as another issue arose:

const source = of(1);
const result = source.pipe(
  switchMap(v => of(v))
);

Problem was on initial subscription chain:
1. result gets subscribed, subscribeStack = [result]
2. source gets subscribed
a) source -> result (OK)
b) subscribeStack = [result, source] (OK, but problematic)
3. source emits, nextStack = [source]
4. switchMap subscribes to the inner observable!
a) inner -> source (ERR!)
b) subscribeStack = [result, source, inner]

I'm not sure why the observable emits before the subscription stack gets popped, but it seems this was the wrong assumption.

Fix was to disable the other stack when next/subscribe happens